REL-4262: migrate to Federated
