+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+##
 ## 0.1.1
 * Fix plugin crashes in background on Android.
 ## 0.1.0
