+++ darwin/Classes/ManagedConfigurationsPlugin.h
