+++ darwin/Classes/ManagedConfigurationsPlugin.m
