+++ darwin/Classes/SwiftManagedConfigurationsPlugin.swift
