+++ pubspec.yaml
@@ -1,28 +1,39 @@
 name: managed_configurations
 description: Plugin to support managed app configuration provided by a Mobile device management (MDM).
-version: 0.1.1
+version: 0.1.2
 homepage: https://github.com/mwaylabs/flutter-managed-configuration
 
+
 environment:
-  sdk: ">=2.12.0"
-  flutter: ">=1.20.0"
+  sdk: ^3.0.0
+  # Flutter versions prior to 3.7 did not support the
+  # sharedDarwinSource option.
+  flutter: ">=3.10.6"
 
 dependencies:
   flutter:
     sdk: flutter
+  plugin_platform_interface: ^2.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  flutter_lints: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
-# The following section is specific to Flutter.
+# The following section is specific to Flutter packages.
 flutter:
   # This section identifies this Flutter project as a plugin project.
-  # The 'pluginClass' and Android 'package' identifiers should not ordinarily
-  # be modified. They are used by the tooling to maintain consistency when
+  # The 'pluginClass' specifies the class (in Java, Kotlin, Swift, Objective-C, etc.)
+  # which should be registered in the plugin registry. This is required for
+  # using method channels.
+  # The Android 'package' specifies package in which the registered class is.
+  # This is required for using method channels on Android.
+  # The 'ffiPlugin' specifies that native code should be built and bundled.
+  # This is required for using `dart:ffi`.
+  # All these are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
   plugin:
     platforms:
@@ -31,4 +42,7 @@ flutter:
         pluginClass: ManagedConfigurationsPlugin
       ios:
         pluginClass: ManagedConfigurationsPlugin
-
+        sharedDarwinSource: true
+      macos:
+        pluginClass: ManagedConfigurationsPlugin
+        sharedDarwinSource: true
